requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.1
pydantic==1.10.2
python-dotenv==0.21.0
